third_party/WebKit/LayoutTests/bluetooth/server/getPrimaryServices/services-not-found.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return getEmptyHealthThermometerDevice()
    .then(([device]) => assert_promise_rejects_with_message(
      device.gatt.getPrimaryServices(),
      new DOMException('No Services found in device.', 'NotFoundError')));
}, 'Request for services in a device with no services. Reject with ' +
   'NotFoundError.');
</script>
